+++ sqlstore/candles.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"code.vegaprotocol.io/data-node/candlesv2"
+	"code.vegaprotocol.io/data-node/metrics"
 
 	"github.com/shopspring/decimal"
 
@@ -77,6 +78,7 @@ func (cs *Candles) GetCandleDataForTimeSpan(ctx context.Context, candleId string
 
 	query, args = orderAndPaginateQuery(query, []string{"period_start"}, p, args...)
 
+	defer metrics.StartSQLQuery("Candles", "GetCandleDataForTimeSpan")()
 	err = pgxscan.Select(ctx, cs.Connection, &candles, query, args...)
 	if err != nil {
 		return nil, fmt.Errorf("querying candles: %w", err)
@@ -121,6 +123,7 @@ func (cs *Candles) GetCandleIdForIntervalAndMarket(ctx context.Context, interval
 func (cs *Candles) getIntervalToView(ctx context.Context) (map[string]string, error) {
 	query := fmt.Sprintf("SELECT view_name FROM timescaledb_information.continuous_aggregates where view_name like '%s%%'",
 		candlesViewNamePrePend)
+	defer metrics.StartSQLQuery("Candles", "GetIntervalToView")()
 	rows, err := cs.Connection.Query(ctx, query)
 	if err != nil {
 		return nil, fmt.Errorf("fetching existing views for interval: %w", err)
@@ -197,6 +200,7 @@ func (cs *Candles) viewExistsForInterval(ctx context.Context, interval string) (
 func (cs *Candles) normaliseInterval(ctx context.Context, interval string) (string, error) {
 	var normalizedInterval string
 
+	defer metrics.StartSQLQuery("Candles", "normaliseInterval")()
 	_, err := cs.Connection.Exec(ctx, "SET intervalstyle = 'postgres_verbose' ")
 	if err != nil {
 		return "", fmt.Errorf("normalising interval, failed to set interval style:%w", err)
@@ -218,6 +222,7 @@ func (cs *Candles) normaliseInterval(ctx context.Context, interval string) (stri
 func (cs *Candles) getIntervalSeconds(ctx context.Context, interval string) (int64, error) {
 	var seconds decimal.Decimal
 
+	defer metrics.StartSQLQuery("Candles", "getIntervalSeconds")()
 	query := fmt.Sprintf("SELECT EXTRACT(epoch FROM INTERVAL '%s')", interval)
 	row := cs.Connection.QueryRow(ctx, query)