+++ sqlstore/epochs.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"code.vegaprotocol.io/data-node/entities"
+	"code.vegaprotocol.io/data-node/metrics"
 	"github.com/georgysavva/scany/pgxscan"
 )
 
@@ -20,6 +21,7 @@ func NewEpochs(connectionSource *ConnectionSource) *Epochs {
 }
 
 func (es *Epochs) Add(ctx context.Context, r entities.Epoch) error {
+	defer metrics.StartSQLQuery("Epochs", "Add")()
 	_, err := es.Connection.Exec(ctx,
 		`INSERT INTO epochs(
 			id,
@@ -38,6 +40,7 @@ func (es *Epochs) Add(ctx context.Context, r entities.Epoch) error {
 }
 
 func (rs *Epochs) GetAll(ctx context.Context) ([]entities.Epoch, error) {
+	defer metrics.StartSQLQuery("Epochs", "GetAll")()
 	epochs := []entities.Epoch{}
 	err := pgxscan.Select(ctx, rs.Connection, &epochs, `
 		SELECT DISTINCT ON (id) * from epochs ORDER BY id, vega_time desc;`)
@@ -45,6 +48,7 @@ func (rs *Epochs) GetAll(ctx context.Context) ([]entities.Epoch, error) {
 }
 
 func (rs *Epochs) Get(ctx context.Context, ID int64) (entities.Epoch, error) {
+	defer metrics.StartSQLQuery("Epochs", "Get")()
 	query := `SELECT DISTINCT ON (id) * FROM epochs WHERE id=$1 ORDER BY id, vega_time desc;`
 
 	epoch := entities.Epoch{}
@@ -59,9 +63,10 @@ func (rs *Epochs) GetCurrent(ctx context.Context) (entities.Epoch, error) {
 	query := `SELECT * FROM epochs ORDER BY id desc, vega_time desc FETCH FIRST ROW ONLY;`
 
 	epoch := entities.Epoch{}
+	defer metrics.StartSQLQuery("Epochs", "GetCurrent")()
 	err := pgxscan.Get(ctx, rs.Connection, &epoch, query)
 	if err != nil {
-		return entities.Epoch{}, fmt.Errorf("querying epocs: %w", err)
+		return entities.Epoch{}, fmt.Errorf("querying epochs: %w", err)
 	}
 	return epoch, nil
 }