+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [526](https://github.com/vegaprotocol/data-node/issues/526) - Add market id and reward type to reward and market to transfer
 - [540](https://github.com/vegaprotocol/data-node/issues/540) - CI: trigger Devnet deployment on merges to develop branch
 - [546](https://github.com/vegaprotocol/data-node/issues/546) - Data retention for margin levels 
+- [578](https://github.com/vegaprotocol/data-node/issues/578) - Add metrics for `SQL` queries
 
 ### 🐛 Fixes
 - [524](https://github.com/vegaprotocol/data-node/issues/524) - Fix for incorrect balances